Moving ColumnList into the features folder
